Moving AuditActivityEnum to audit.api package
- also renaming to `AuditActivity`
