+++ src/main/java/uk/gov/hmcts/darts/audit/api/AuditApi.java
@@ -1,10 +1,9 @@
 package uk.gov.hmcts.darts.audit.api;
 
-import uk.gov.hmcts.darts.audit.enums.AuditActivityEnum;
 import uk.gov.hmcts.darts.common.entity.CourtCaseEntity;
 import uk.gov.hmcts.darts.common.entity.UserAccountEntity;
 
 public interface AuditApi {
 
-    void recordAudit(AuditActivityEnum activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase);
+    void recordAudit(AuditActivity activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase);
 }
