+++ src/main/java/uk/gov/hmcts/darts/audit/api/impl/AuditApiImpl.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.audit.api.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.darts.audit.api.AuditApi;
-import uk.gov.hmcts.darts.audit.enums.AuditActivityEnum;
+import uk.gov.hmcts.darts.audit.api.AuditActivity;
 import uk.gov.hmcts.darts.audit.service.AuditService;
 import uk.gov.hmcts.darts.common.entity.CourtCaseEntity;
 import uk.gov.hmcts.darts.common.entity.UserAccountEntity;
@@ -15,7 +15,7 @@ public class AuditApiImpl implements AuditApi {
     private final AuditService auditService;
 
     @Override
-    public void recordAudit(AuditActivityEnum activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase) {
+    public void recordAudit(AuditActivity activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase) {
         auditService.recordAudit(activity, userAccountEntity, courtCase);
     }
 }
