+++ src/main/java/uk/gov/hmcts/darts/audit/service/AuditService.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.darts.audit.service;
 
 import jakarta.transaction.Transactional;
-import uk.gov.hmcts.darts.audit.enums.AuditActivityEnum;
+import uk.gov.hmcts.darts.audit.api.AuditActivity;
 import uk.gov.hmcts.darts.audit.model.AuditSearchQuery;
 import uk.gov.hmcts.darts.common.entity.AuditEntity;
 import uk.gov.hmcts.darts.common.entity.CourtCaseEntity;
@@ -12,7 +12,7 @@ import java.util.List;
 public interface AuditService {
 
     @Transactional
-    void recordAudit(AuditActivityEnum activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase);
+    void recordAudit(AuditActivity activity, UserAccountEntity userAccountEntity, CourtCaseEntity courtCase);
 
     List<AuditEntity> search(AuditSearchQuery auditSearchQuery);
 }
